requirements.txt
django==5.2.1
opencv-python==4.10.0.84
ultralytics==8.3.50
numpy==2.1.3
psycopg2-binary==2.9.10
cv2-enumerate-cameras==1.3.3
huggingface-hub==0.33.0
